+++ lua/cold_word_drop/hide_words.lua
@@ -1,4 +1,4 @@
 local hide_words =
-{ 	["示~例~"] = { "shil", "shili", },
+{ ["示~例~"] = { "shil", "shili", },
 }
 return hide_words