+++ lua/cold_word_drop/turndown_freq_words.lua
@@ -1,4 +1,4 @@
 local turndown_freq_words =
-{ 	["示~例~"] = { "shili", },
+{ ["示~例~"] = { "shili", },
 }
 return turndown_freq_words