+++ rime.lua
@@ -35,30 +35,16 @@ autocap_filter = require("autocap_filter")
 -- 降低部分英语单词在候选项的位置，可在方案中配置要降低的模式和单词
 reduce_english_filter = require("reduce_english_filter")
 
--- 默认未启用：
+-- 辅码，https://github.com/mirtlecn/rime-radical-pinyin/blob/master/search.lua.md
+search = require("search")
 
--- 词条置顶
--- 满足左边的 cand.preedit 时，将右边的 cand 按顺序置顶。
--- 在 engine/filters 增加 - lua_filter@pin_cand_filter
--- 在方案里写配置项：
--- pin_cand_filter:
---   - "l	了"
---   - "le	了"
---   - "ta	他 她 它"
---   - "ni hao	你好 拟好"
+-- 置顶候选项
 pin_cand_filter = require("pin_cand_filter")
 
 -- 长词优先（全拼）
--- 在 engine/filters 增加 - lua_filter@long_word_filter
--- 在方案里写配置项:
--- 提升 count 个词语，插入到第 idx 个位置。
--- 示例：将 2 个词插入到第 4、5 个候选项，输入 jie 得到「1接 2解 3姐 4饥饿 5极恶」
--- long_word_filter:
---   count: 2
---   idx: 4
--- 使用请注意：之前有较多网友反应有内存泄漏，优化过一些但还是偶尔有较高的内存，但并不卡顿也不影响性能，重新部署后即正常
--- 如果要启用，建议放到靠后位置，最后一个放 uniquifier，倒数第二个就放 long_word_filter
 long_word_filter = require("long_word_filter")
+
+-- 默认未启用：
 
 -- 中英混输词条自动空格
 -- 在 engine/filters 增加 - lua_filter@cn_en_spacer
@@ -88,3 +74,24 @@ is_in_user_dict = require("is_in_user_dict")
 -- get_record_filername() 函数中仅支持了 Windows、macOS、Linux
 cold_word_drop_processor = require("cold_word_drop.processor")
 cold_word_drop_filter = require("cold_word_drop.filter")
+
+
+-- 暴力 GC
+-- 详情 https://github.com/hchunhui/librime-lua/issues/307
+-- 这样也不会导致卡顿，那就每次都调用一下吧，内存稳稳的
+function force_gc()
+    -- collectgarbage()
+    collectgarbage("step")
+end
+
+-- 临时用的
+function debug_checker(input, env)
+    for cand in input:iter() do
+        yield(ShadowCandidate(
+            cand,
+            cand.type,
+            cand.text,
+            env.engine.context.input .. " - " .. env.engine.context:get_preedit().text .. " - " .. cand.preedit
+        ))
+    end
+end