+++ others/CHANGELOG.md
@@ -4,6 +4,33 @@
 
 
 
+## 2024-02-09 ♻️
+
+重构了 `pin_cand_filter.lua` 置顶候选项功能。 [#675](https://github.com/iDvel/rime-ice/issues/675)
+
+- 调整了方案中 `engine/filters` 的排序
+- 不再需要在配置中写 emoji，emoji 可自动吸附。
+
+## 2024-02-04 ⚠️
+
+⚠️ 中英混输词库由英文方案附属切换到单独的 table_translator ([#662](https://github.com/iDvel/rime-ice/pull/662))
+
+- 词库文件由 `cn_en*.dict.yaml` 变为 `cn_en*.txt`
+- 双拼不再需要去 `melt_eng.dict.yaml` 更改引用词库
+
+## 2024-02-01 🆕
+
+`pin_cand_filter.lua` 置顶候选项 [#586](https://github.com/iDvel/rime-ice/issues/586)
+
+## 2024-01-29 🆕 ⚠️
+
+[部件拆字方案](https://github.com/mirtlecn/rime-radical-pinyin) 反查、辅码 ([#643](https://github.com/iDvel/rime-ice/pull/643))
+
+- 反查：默认以 `uU` 开头
+- 辅码：默认以 `` ` ``（反引号）开启查询
+
+⚠️ 部件拆字方案替换掉了两分方案
+
 ## 2024-01-02 🆕 🐛 ⚠️
 
 🆕 农历功能 [#565](https://github.com/iDvel/rime-ice/issues/565)
@@ -12,7 +39,7 @@
 
 ⚠️ 更新并修改 `weasel_style.yaml` 为 `weasel.yaml` （[#584](https://github.com/iDvel/rime-ice/pull/584)）
 
-## 2023-11-29
+## 2023-11-29 ⚠️
 
 九宫格方案 2~9 的映射由 ADGJMPTW 改为 23456789 [a0e0ef8](https://github.com/iDvel/rime-ice/commit/a0e0ef807e4ebc50771563717375f554c9473315)
 
@@ -69,9 +96,9 @@
 - 增加 Lua 脚本 `autocap_filter.lua`
 - 修改了 `melt_eng.schema.yaml` 拼写派生规则。
 
-## 2023-06-15 🐛
+## 2023-05-24 🐛
 
-修复全拼方案模糊音派生规则
+修复全拼方案模糊音派生规则 [6c0618a](https://github.com/iDvel/rime-ice/commit/6c0618aeaf2910482e20ff1c057f482aaa98c051)
 
 ## 2023-05-13 🆕